Rename gmp python module to xml
The module is responsible to create gmp xml commands. In future this
file may be splittet into xml and gmp command related purposes.
