+++ gmp/xml.py
